+++ lib/socialization/lib/exceptions.rb
@@ -1,0 +1,3 @@
+module Socialization
+  class ArgumentError < ::ArgumentError; end
+end